Version 1.3a
MeetingsApp
- Added the link to end currently running meetings

Videoserver
- Added the functionality to end a particular meeting
- Reorganized the server js into different routes
- Created a new page for the delete link
