+++ .gitignore
@@ -13,3 +13,4 @@ Transcription/transcription_env/
 videoserver/config/config.js
 MaxLabProject/MaxLabProject/settings_config.py
 Backup/
+LocalRecordings/
