+++ videoserver/views/InvalidMeeting.ejs
@@ -1,36 +1,45 @@
 <html>
-   <head>
-      <title>MaxLab Meetings</title>
-      <meta name="viewport" content="width=device-width, initial-scale=1" charset="utf-8">
-      <!-- Bootstrap -->
-      <link rel="stylesheet" href=
-         "https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css">
-      <script src=
-         "https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
-      <script src=
-         "https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"></script>
-      <script src=
-         "https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
-      <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-      <!-- Bootstrap -->
-      <script defer>
-        const MESSAGE = "<%= data%>"
-        console.error(MESSAGE)
-       
-     </script>
-   </head>
-   <body>
-    <div id="invalidMeeting" style="display: block" >
-        <div class="border d-flex flex-column min-vh-100 justify-content-center align-items-center ">
-           <div class="jumbotron w-50 flex-grow-0 flex-shrink-1" id="gateway">
-              <h1 class="text-center display-3">Error!</h1>
-              <p id = "message" class=" text-center display-4"></p>
-           </div>
+    <head>
+        <title>MaxLab Meetings</title>
+        <meta
+            name="viewport"
+            content="width=device-width, initial-scale=1"
+            charset="utf-8"
+        />
+        <!-- Bootstrap -->
+        <link
+            rel="stylesheet"
+            href="https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
+        />
+        <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+        <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"></script>
+        <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
+        <link
+            rel="stylesheet"
+            href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css"
+        />
+        <!-- Bootstrap -->
+        <script defer>
+            const MESSAGE = "<%= data%>";
+            console.error(MESSAGE);
+        </script>
+    </head>
+    <body>
+        <div id="invalidMeeting" style="display: block">
+            <div
+                class="border d-flex flex-column min-vh-100 justify-content-center align-items-center"
+            >
+                <div
+                    class="jumbotron w-50 flex-grow-0 flex-shrink-1"
+                    id="gateway"
+                >
+                    <h1 class="text-center display-3">Error!</h1>
+                    <p id="message" class="text-center display-4"></p>
+                </div>
+            </div>
         </div>
-     </div>
-            
-   </body>
-   <script>
-        document.getElementById("message").innerHTML=MESSAGE;        
-   </script>
+    </body>
+    <script>
+        document.getElementById("message").innerHTML = MESSAGE;
+    </script>
 </html>
